Move extrapolation methods to an add-on library
This is done in a non-breaking way with backwards compatibility to start so we can gather all of these changes and do all of the move-outs all at once.
