+++ src/alg_utils.jl
@@ -408,29 +408,6 @@ get_current_alg_order(alg::QNDF, cache) = cache.order
 get_current_alg_order(alg::FBDF, cache) = cache.order
 get_current_adaptive_order(alg::QNDF, cache) = cache.order
 get_current_adaptive_order(alg::FBDF, cache) = cache.order
-function get_current_adaptive_order(
-        alg::OrdinaryDiffEqExtrapolationVarOrderVarStepAlgorithm,
-        cache)
-    cache.cur_order
-end
-function get_current_alg_order(alg::OrdinaryDiffEqExtrapolationVarOrderVarStepAlgorithm,
-        cache)
-    cache.cur_order
-end
-get_current_alg_order(alg::ExtrapolationMidpointDeuflhard, cache) = 2(cache.n_curr + 1)
-get_current_alg_order(alg::ImplicitDeuflhardExtrapolation, cache) = 2(cache.n_curr + 1)
-get_current_adaptive_order(alg::ExtrapolationMidpointDeuflhard, cache) = 2cache.n_curr
-get_current_adaptive_order(alg::ImplicitDeuflhardExtrapolation, cache) = 2cache.n_curr
-get_current_alg_order(alg::ExtrapolationMidpointHairerWanner, cache) = 2(cache.n_curr + 1)
-get_current_alg_order(alg::ImplicitHairerWannerExtrapolation, cache) = 2(cache.n_curr + 1)
-get_current_alg_order(alg::ImplicitEulerBarycentricExtrapolation, cache) = cache.n_curr
-get_current_alg_order(alg::ImplicitEulerExtrapolation, cache) = cache.n_curr + 1
-get_current_adaptive_order(alg::ExtrapolationMidpointHairerWanner, cache) = 2cache.n_curr
-get_current_adaptive_order(alg::ImplicitHairerWannerExtrapolation, cache) = 2cache.n_curr
-get_current_adaptive_order(alg::ImplicitEulerExtrapolation, cache) = cache.n_curr - 1
-function get_current_adaptive_order(alg::ImplicitEulerBarycentricExtrapolation, cache)
-    cache.n_curr - 2
-end
 
 #alg_adaptive_order(alg::OrdinaryDiffEqAdaptiveAlgorithm) = error("Algorithm is adaptive with no order")
 function get_current_adaptive_order(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm},
@@ -443,7 +420,6 @@ end
 
 alg_order(alg::FunctionMap) = 0
 alg_order(alg::Euler) = 1
-alg_order(alg::AitkenNeville) = alg.init_order
 alg_order(alg::Heun) = 2
 alg_order(alg::Ralston) = 2
 alg_order(alg::LawsonEuler) = 1
@@ -740,13 +716,7 @@ alg_order(alg::QPRK98) = 9
 
 alg_maximum_order(alg) = alg_order(alg)
 alg_maximum_order(alg::CompositeAlgorithm) = maximum(alg_order(x) for x in alg.algs)
-alg_maximum_order(alg::ExtrapolationMidpointDeuflhard) = 2(alg.max_order + 1)
-alg_maximum_order(alg::ImplicitDeuflhardExtrapolation) = 2(alg.max_order + 1)
-alg_maximum_order(alg::ExtrapolationMidpointHairerWanner) = 2(alg.max_order + 1)
-alg_maximum_order(alg::ImplicitHairerWannerExtrapolation) = 2(alg.max_order + 1)
-alg_maximum_order(alg::ImplicitEulerExtrapolation) = 2(alg.max_order + 1)
 
-alg_maximum_order(alg::ImplicitEulerBarycentricExtrapolation) = alg.max_order
 alg_adaptive_order(alg::ExplicitRK) = alg.tableau.adaptiveorder
 alg_adaptive_order(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm}) = alg_order(alg) - 1
 alg_adaptive_order(alg::Feagin10) = 8
@@ -782,20 +752,6 @@ function default_controller(alg, cache, qoldinit, _beta1 = nothing, _beta2 = not
         beta1, beta2 = _digest_beta1_beta2(alg, cache, Val(QT), _beta1, _beta2)
         return PIController(beta1, beta2)
     end
-end
-
-function default_controller(
-        alg::Union{ExtrapolationMidpointDeuflhard,
-            ImplicitDeuflhardExtrapolation,
-            ExtrapolationMidpointHairerWanner,
-            ImplicitHairerWannerExtrapolation,
-            ImplicitEulerExtrapolation,
-            ImplicitEulerBarycentricExtrapolation},
-        cache,
-        qoldinit, _beta1 = nothing, _beta2 = nothing)
-    QT = typeof(qoldinit)
-    beta1, beta2 = _digest_beta1_beta2(alg, cache, Val(QT), _beta1, _beta2)
-    return ExtrapolationController(beta1)
 end
 
 function _digest_beta1_beta2(alg, cache, ::Val{QT}, _beta1, _beta2) where {QT}
@@ -852,12 +808,6 @@ end
 beta2_default(alg::FunctionMap) = 0
 beta2_default(alg::DP8) = 0 // 1
 beta2_default(alg::DP5) = 4 // 100
-beta2_default(alg::ExtrapolationMidpointDeuflhard) = 0 // 1
-beta2_default(alg::ImplicitDeuflhardExtrapolation) = 0 // 1
-beta2_default(alg::ExtrapolationMidpointHairerWanner) = 0 // 1
-beta2_default(alg::ImplicitHairerWannerExtrapolation) = 0 // 1
-beta2_default(alg::ImplicitEulerExtrapolation) = 0 // 1
-beta2_default(alg::ImplicitEulerBarycentricExtrapolation) = 0 // 1
 
 function beta1_default(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm}, beta2)
     isadaptive(alg) ? 7 // (10alg_order(alg)) : 0
@@ -865,12 +815,6 @@ end
 beta1_default(alg::FunctionMap, beta2) = 0
 beta1_default(alg::DP8, beta2) = typeof(beta2)(1 // alg_order(alg)) - beta2 / 5
 beta1_default(alg::DP5, beta2) = typeof(beta2)(1 // alg_order(alg)) - 3beta2 / 4
-beta1_default(alg::ExtrapolationMidpointDeuflhard, beta2) = 1 // (2alg.init_order + 1)
-beta1_default(alg::ImplicitDeuflhardExtrapolation, beta2) = 1 // (2alg.init_order + 1)
-beta1_default(alg::ExtrapolationMidpointHairerWanner, beta2) = 1 // (2alg.init_order + 1)
-beta1_default(alg::ImplicitHairerWannerExtrapolation, beta2) = 1 // (2alg.init_order + 1)
-beta1_default(alg::ImplicitEulerExtrapolation, beta2) = 1 // (alg.init_order + 1)
-beta1_default(alg::ImplicitEulerBarycentricExtrapolation, beta2) = 1 // (alg.init_order - 1)
 
 function gamma_default(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm})
     isadaptive(alg) ? 9 // 10 : 0
@@ -878,25 +822,6 @@ end
 gamma_default(alg::CompositeAlgorithm) = maximum(gamma_default, alg.algs)
 gamma_default(alg::RKC) = 8 // 10
 gamma_default(alg::IRKC) = 8 // 10
-function gamma_default(alg::ExtrapolationMidpointDeuflhard)
-    (1 // 4)^beta1_default(alg, beta2_default(alg))
-end
-function gamma_default(alg::ImplicitDeuflhardExtrapolation)
-    (1 // 4)^beta1_default(alg, beta2_default(alg))
-end
-function gamma_default(alg::ExtrapolationMidpointHairerWanner)
-    (65 // 100)^beta1_default(alg, beta2_default(alg))
-end
-function gamma_default(alg::ImplicitHairerWannerExtrapolation)
-    (65 // 100)^beta1_default(alg, beta2_default(alg))
-end
-function gamma_default(alg::ImplicitEulerExtrapolation)
-    (65 // 100)^beta1_default(alg, beta2_default(alg))
-end
-
-function gamma_default(alg::ImplicitEulerBarycentricExtrapolation)
-    (80 // 100)^beta1_default(alg, beta2_default(alg))
-end
 
 qsteady_min_default(alg::Union{OrdinaryDiffEqAlgorithm, DAEAlgorithm}) = 1
 qsteady_min_default(alg::FBDF) = 9 // 10
